backend/requirements.txt
djangorestframework
djangorestframework-simplejwt
Django
psycopg2-binary==2.9.3
gunicorn
Pillow
django-phonenumber-field
phonenumbers
